Function Debug
The Story
Today the students and TA's have flipped roles! We wrote some functions, and you have to tell us if they are good or if they have bugs.

The Problem
You will have to come up with test cases for seven given functions which we describe below. You can run the code locally as many times as you want passing as many arguments as you would like into our functions. The actual code that you will be submitting is in the potd.cpp file. In this file there is a function called evaluate. All that this function should do is return a vector of integers. Vector index i should be 0 if the ith function is has a bug and 1 if the function is correct.

The functions' behavior is outlined in the Functions.h file. Note that it is "proper behavior" if our functions print out an error message and exit because the input to the function was outside of the range specified in the specs (listed in the Functions.h file).

Important
As there are only 7 functions, and we want to prevent guessing and checking, you will only get 3 submissions for this potd. After that, you will no longer be able to submit.

Go forth and test!