Hash: Function and String Order
The Problem
Task 1: Implement the Bernstein hash function
Today, we’ll construct a more sophisticated hash function, called the Bernstein hash. For a string str, the Bernstein hash, b_hash, is computed by iterating over its characters. The pseudo-code is as shown below:

    b_hash = 5381
    for all character in str:
        b_hash *= 33 + character //use the ASCII value of the character here
The magic number 33 is used simply because it works! Why it works better than many other constants, prime or not, has never been adequately explained. Your task is to write a function unsigned long int bernstein(string str) that takes in a string as an input and returns its hash. Notice how quickly the hash grows with the length of the input string.

Task 2: Effect of character ordering on the Bernstein hash
We’ll now test the hash function with various inputs. Recall that the additive hash (the one you implemented in the last POTD) is invariant to the ordering of the letters in a string. Thus, POTS and STOP have the same additive hash. This is usually not desirable. We’d first like to see if the Bernstein hash function is sensitive to the ordering of characters. To do so, we’ll compare the hashes of a string and its reverse. Write a function reverse() to invert a string, and check if the Bernstein hash for a string changes if the string is inverted.

Example Output:
Bernstein hash of POTS with range 13 is: 6
Bernstein hash of the reverse of POTS - STOP - with range 13, is: 7
Compile and Test
A complete Makefile and a main.cpp file containing one simple test has been provided for you. To compile and run, run:

make
./main