Hash Map: add up values with commen keys
The Problem

In lecture, we learned about hash table and how it is implemented. In today's POTD, we are going to practice how to use unordered_map, the hash table in c++ standard library. Finish the function common_elems which takes in 2 unordered_map references as parameters. It is supposed to take out all the common keys from mapA and mapB, and return a map containing the sum of their corresponding values.

You can read about unordered_map here.
Example

Assume mapA contains {"a", 1}, {"b", 2}, {"c", 3} and mapB contains {"a", 4}, {"b", 5}, {"d", 6}, we should return a map containing {"a", 5}, {"b", 7} and mapA should be left with {"c", 3}, and mapB with {"d", 6}.