Hash: table, saperate chain
The Problem
Today’s task is to implement some functions for a hash table that utilizes the separate chaining for collision resolution. You are given the HashTable class containing the required data variables and the constructor. The hash table itself is stored as an array of vectors, called table. The hashing function used is the Bernstein hash. It has already been implemented for you. You must implement the insert() and contains() functions. Note that the insert function must not allow duplicate keys to be inserted into the table.

Example Output:
Does the hash table contain bbb? 0
Does the hash table contain aaa? 1
Contents of the hash table:
ccc
aaa
Compile and Test
A complete Makefile and a main.cpp file containing one simple test has been provided for you.