Dictionary: Map, sort by the frequency of words
The Problem
Today we have a little exercise to help you get used to working with dictionaries. Though you do not have to use a dictionary to solve this problem, it is recommended you do.

You have intercepted an encrypted message sent from Wade to Taylor Swift. Luckily, you overheard Wade talking to her on the phone, and know how to decrypt these messages they send. The encryption scheme is as follows:

Think of each word in the message as having an index, starting at index one. Repeat the word at index i i-times in the encrypted message. Then suffle the words in the encrypted message.

For example, if the original message was: "I like cats", the encrypted message could be "like I cats cats like cats", or any other permutation of the order of words (i.e. "cats I cats like like cats" is another valid encryption of "I like cats"). This is not the best encryption scheme, as you cannot have repeated words in the message, but when Taylor Swift gives you an encryption scheme, you don't question it.

Your job is to write the function decipherer which decrypts a message encrypted with the scheme described above. The function takes one parameter file, the file name of the encrypted message. This file has one word per line and the frequency of each word in the file is unique. The function should return the decrypted message as a string with spaces separating each word in the message. There should be no space at the beginning of the message or after the last word in the message.